+++ src/components/Modal/CommonModalStyles.ts
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+export const ModalDescription = styled.p`
+  font-size: 1.2rem;
+  text-align: center;
+`;
+
+export const ModalButtonContainer = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: space-evenly;
+`;
+
+export const ModalButton = styled.button`
+  width: 8rem;
+  height: 2.5rem;
+  border-radius: 8px;
+  background-color: rgba(100, 212, 171, 1);
+  color: white;
+  line-height: 10%;
+`;
